Major commit with report and some final changes

+++ project2/code/main.py
@@ -23,7 +23,8 @@ def delete_files(paths, kwargs=0):
         if not isinstance(kwargs, list):
             kwargs = [kwargs]
         for _file in kwargs:
-            os.remove(paths + _file)
+            if (os.path.isfile(paths + _file)):
+                os.remove(paths + _file)
         if not os.listdir(paths):
             os.rmdir(paths)
 
@@ -84,7 +85,7 @@ if problem == "bb":
     print("BUCKLING BEAM \n")
 
     # Create directories for output if not already exsisting
-    create_directory([path_data, path_plot, path_data_output])
+    create_directory([path_data, path_plot, path_data_output, path_GEN])
 
     # Remove files containing iterations if they exsist.
     for string in ["bi_", "ja_"]:
@@ -139,6 +140,7 @@ if problem == "bb":
         else:
             k = int(k)
         for j in range(k):
+            print(f"Running for k = {j+1:d}")
             for i in m:
                 system(path_cpp + " " + str(i) +
                        " 0 " + filename_data + " 0 1 " + tolerance + " 2")
@@ -149,7 +151,7 @@ if problem == "bb":
             system("python3 " + path_scripts +
                    "plot_timecomp.py " + filename_data)
     else:
-        print("Do you already have a benchmark you want to plot? Default: n")
+        print("Do you already have a benchmark you want to plot? Default: no")
         plot_timecomp = input("[y/n] ")
         if (plot_timecomp == "y"):
             system("python3 " + path_scripts +
@@ -180,6 +182,10 @@ if problem == "bb":
         system("python3 " + path_scripts + "plot_eigenvector.py " +
                filename_data + " " + n + " " + largest_eigenvalue)
 
+    if (os.path.isfile(path_data_ouput + "data_BB" + n)):
+        system("mv " + path_data_output + "data_BB" +
+               n + " " + path_data + "/data_BB" + n)
+
     print("Do you want to delete the output files? Default: yes")
     delete = input("[y/n] ")
 
@@ -202,11 +208,16 @@ if problem == "bb":
         if (delete != "n"):
             delete_files(path_data_output)
 
-    print("Do you want to delete general files? Default: no")
-    print("These files are in the folder GEN_data, and are the files containing time comparison and iterations.")
-    delete = input("[y/n] ")
-    if (delete == "y"):
-        delete_files(path_GEN)
+    # print("Do you want to delete general files? Default: no")
+    # print("These files are in the folder GEN_data, and are the files containing time comparison and iterations.")
+    # delete = input("[y/n] ")
+    # if (delete == "y"):
+    #     dirs = os.listdir(path_GEN)
+    #     directory = []
+    #     for _dir in dirs:
+    #         if ("BB" in _dir):
+    #             directory.append(_dir)
+    #     delete_files(path_GEN, directory)
 
     print("Thank you... Have a nice day.")
 
@@ -220,7 +231,7 @@ if problem == "qd1e":
     print("QUANTUM DOTS ONE ELECTRON \n")
 
     # Create directories for output if not already exsisting
-    create_directory([path_data, path_plot, path_data_output])
+    create_directory([path_data, path_data_output, path_GEN])
 
     print("Enter desired tolerance: Default: 10^(-5)")
     tolerance = input("10^(-[int]) ")
@@ -280,16 +291,29 @@ if problem == "qd1e":
             for n in multi_n:
                 # "./cpp_codes/main {n} 1 {filename_data}{rho_max} 0 {rho_max} {tolerance} 1"
                 system(path_cpp + " " + str(n) + " 1 " +
-                       filename_data + str(rho_names[i]) + " 0 " + str(rho) + " " + tolerance + " 1")
+                       filename_data + str(rho_names[i]) + "_" + " 0 " + str(rho) + " " + tolerance + " 1")
 
     print("Finding the smallest values for rho and n, which satisfies the condition of the tolerance, if not satisfied, returns the smallest error.")
-    system("python3 ./scipts/eigenvalue_error.py " +
+    system("python3 " + path_scripts + "eigenvalues_error.py " +
            filename_data + " " + tolerance)
+
+    move_list = []
+    with open(path_GEN + "eigenvalues_rho_n", 'r') as f:
+        for lines in f:
+            lines = lines.strip("(")
+            lines = lines.strip(")\n")
+            lines = lines.split(",")
+            line = lines[-1]
+            move_list.append(line.strip(" "))
+    for string in move_list:
+        if (os.path.isfile(path_data_output + "data_" + filename_data + string)):
+            system("mv " + path_data_output + "data_" + filename_data +
+                   string + " " + path_data + "/data_" + filename_data + string)
 
     print("Do you want to delete the output files? Default: yes")
     delete = input("[y/n] ")
 
-    if (delete == "y"):
+    if (delete != "n"):
         # Asking if all files should be deleted or only spesific
         files = []
         print("Specify which files, if NONE, all are deleted: Default: NONE")
@@ -302,16 +326,17 @@ if problem == "qd1e":
             delete_files(path_data_output)
         else:
             delete_files(path_data_output, files)
-    elif(delete != "n"):
-        print("You sure you want to delete the files? Default: yes")
-        delete = input("[y/n]  ")
-        if (delete != "n"):
-            delete_files(path_data_output)
 
     print("Do you want to delete general files? Default: no")
     print("These files are in the folder GEN_data, and are the files containing time comparison and iterations.")
     delete = input("[y/n] ")
     if (delete == "y"):
-        delete_files(path_GEN)
+        dirs = os.listdir(path_GEN)
+        directory = []
+        for _dir in dirs:
+            if ("QD" in _dir):
+                directory.append(_dir)
+            if directory:
+                delete_files(path_GEN, directory)
 
     print("Thank you... Have a nice day.")
